requirements.txt
streamlit==1.32.0
torch
transformers
sentence-transformers
faiss-cpu
accelerate
pandas